+++ spi.h
@@ -22,5 +22,5 @@ void initSPI(void);
 *@param[in] csPin The GPIO pin that the CS line is connected to for SPI1
 *@retval None
 */
-void SPI_DMA_Transfer(const uint8_t* rx, const uint8_t* tx, const uint8_t bufferSize, GPIO_TypeDef* csPort, uint8_t csPin);
+void SPI_DMA_Transfer(const uint8_t* rx, const uint8_t* tx, const uint8_t bufferSize, GPIO_TypeDef* csPort, uint16_t csPin);
 #endif
